+++ src/Icon/Icon.js
@@ -5,11 +5,11 @@ import withTheme from '../withTheme'
 /**
  * ⚠️ Requires SVGs are inlined into document somewhere
  */
-export default withTheme('Icon', ({ color, id, font = true, ...rest }) =>
+export default withTheme('Icon', ({ id, font = true, ...rest }) =>
   elem({
     defaultAs: 'svg',
     role: 'img',
-    classes: [`icon icon-${id}`, { font, [`text-${color}`]: color }],
+    classes: [`icon icon-${id}`, { font }],
     children: <use href={`#${id}`} xlinkHref={`#${id}`} />,
     ...rest,
   }),